Make it all user-settable.
