+++ Sources/Segment/Configuration.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import JSONSafeEncoder
 #if os(Linux)
 import FoundationNetworking
 #endif
@@ -37,10 +38,10 @@ public class Configuration {
         var requestFactory: ((URLRequest) -> URLRequest)? = nil
         var errorHandler: ((Error) -> Void)? = nil
         var flushPolicies: [FlushPolicy] = [CountBasedFlushPolicy(), IntervalBasedFlushPolicy()]
-
         var operatingMode: OperatingMode = .asynchronous
         var flushQueue: DispatchQueue = OperatingMode.defaultQueue
         var userAgent: String? = nil
+        var jsonNonConformingNumberStrategy: JSONSafeEncoder.NonConformingFloatEncodingStrategy = .zero
     }
     
     internal var values: Values
@@ -216,9 +217,19 @@ public extension Configuration {
         return self
     }
 
+    /// Specify a custom UserAgent string.  This bypasses the OS dependent check entirely.
     @discardableResult
     func userAgent(_ userAgent: String) -> Configuration {
         values.userAgent = userAgent
+        return self
+    }
+    
+    /// This option specifies how NaN/Infinity are handled when encoding JSON.
+    /// The default is .zero.  See JSONSafeEncoder.NonConformingFloatEncodingStrategy for more informatino.
+    @discardableResult
+    func jsonNonConformingNumberStrategy(_ strategy: JSONSafeEncoder.NonConformingFloatEncodingStrategy) -> Configuration {
+        values.jsonNonConformingNumberStrategy = strategy
+        JSON.jsonNonConformingNumberStrategy = values.jsonNonConformingNumberStrategy
         return self
     }
 }
